Moved out request logic into separate module.
