+++ src/services/appiumService.js
@@ -4,56 +4,7 @@ const { AppiumError } = require("../errors");
 const { NotImplementedError } = require("../errors");
 const { platform, isInstanceOf, getRelativePoint, toBoolean, toNumber } = require("../utils");
 const { transformBounds } = require("../attributeTransforms");
-
-const BASE_URL = configService.getBaseUrl();
-
-const qs = (params) => {
-  if (!params) {
-    return "";
-  }
-
-  const pairs = Object.keys(params)
-    .reduce((result, key) => {
-      const value = params[key];
-
-      if (value === null || typeof value === "undefined") {
-        return result;
-      }
-
-      return result.concat(`${key}=${encodeURIComponent(value)}`);
-    }, []);
-
-  if (!pairs.length) {
-    return "";
-  }
-
-  return "?" + pairs.join("&");
-};
-
-const request = ({ method, path, query, payload, transform }) => {
-  const queryString = qs(query);
-  const opts = {
-    method: method,
-    body: payload
-      ? JSON.stringify(payload)
-      : undefined,
-    headers: {
-      "Content-Type": "application/json"
-    }
-  };
-
-  return fetch(`${BASE_URL}${path}${queryString}`, opts)
-    .then((res) => res.json())
-    .then((data) => {
-      if (data.status) {
-        throw new AppiumError(data.value.message, data.status);
-      }
-
-      return transform
-        ? transform(data)
-        : data.value;
-    });
-};
+const request = require("./request");
 
 const parseValue = (rawValue, elementType, options) => {
   switch (elementType) {
